aidl: fingerprint: Fix UDFPS enrollment and HBM lifecycle on A52s
Enrollment on the Egistec ET713 optical sensor never completed: BAuth
accepted ~4 samples, then rejected every capture as PARTIAL (TEE ret 39)
until it aborted with ret 70.

Manage display HBM per touch and release it whenever an operation ends:

- onPointerDown/onPointerUp write mask_brightness (331/0) and wait for
  actual_mask_brightness via sysfs poll, so a capture never races a
  still-settling panel. HBM must NOT be held for a whole enrollment
  session; BAuth needs the illumination to cycle between captures.
- Release HBM in the FINGERPRINT_AUTHENTICATED handler and on enrollment
  completion. A successful unlock tears down the UDFPS overlay before
  onPointerUp reaches the HAL, so relying on finger-up alone left the
  panel pinned at 547 nits and made brightness controls appear dead.
- Create the BAuth user template directory and the /data/vendor/biometrics
  tree, which Samsung's framework creates on OneUI but AOSP does not.
- Clear stale finger_mask state at HAL start so the first HBM write is
  not swallowed by the kernel's "mask already enabled" state check.
- Log SEM capture events and surface CAPTURE_FAILED to the framework.
- Do not close the BAuth TEE session in Session::close(); the framework
  recreates Sessions and closing invalidates the challenge (error 61).
- Build with -Werror and conversion warnings; fix the resulting casts.

Enrollment now runs to completion (17 accepts / 2 rejects) and unlock,
AOD unlock and brightness control all work.

+++ aidl/fingerprint/Legacy2Aidl.h
@@ -21,23 +21,23 @@ namespace biometrics {
 namespace fingerprint {
 
 inline void translate(const HardwareAuthToken& authToken, hw_auth_token_t& hat) {
-    hat.challenge = authToken.challenge;
-    hat.user_id = authToken.userId;
-    hat.authenticator_id = authToken.authenticatorId;
+    hat.challenge = static_cast<uint64_t>(authToken.challenge);
+    hat.user_id = static_cast<uint64_t>(authToken.userId);
+    hat.authenticator_id = static_cast<uint64_t>(authToken.authenticatorId);
     // these are in host order: translate to network order
     hat.authenticator_type = htobe32(static_cast<uint32_t>(authToken.authenticatorType));
-    hat.timestamp = htobe64(authToken.timestamp.milliSeconds);
+    hat.timestamp = htobe64(static_cast<uint64_t>(authToken.timestamp.milliSeconds));
     std::copy(authToken.mac.begin(), authToken.mac.end(), hat.hmac);
 }
 
 inline void translate(const hw_auth_token_t& hat, HardwareAuthToken& authToken) {
-    authToken.challenge = hat.challenge;
-    authToken.userId = hat.user_id;
-    authToken.authenticatorId = hat.authenticator_id;
+    authToken.challenge = static_cast<int64_t>(hat.challenge);
+    authToken.userId = static_cast<int64_t>(hat.user_id);
+    authToken.authenticatorId = static_cast<int64_t>(hat.authenticator_id);
     // these are in network order: translate to host
     authToken.authenticatorType =
             static_cast<keymaster::HardwareAuthenticatorType>(be32toh(hat.authenticator_type));
-    authToken.timestamp.milliSeconds = be64toh(hat.timestamp);
+    authToken.timestamp.milliSeconds = static_cast<int64_t>(be64toh(hat.timestamp));
     authToken.mac.insert(authToken.mac.begin(), std::begin(hat.hmac), std::end(hat.hmac));
 }
